Remove custom GDT & memory utilities, replace with KLibC library
In this commit, we have removed the custom implementation of GlobalDescriptorTable, interrupt handling and memory utilities. This is replaced with a simplified KLibC library which contains a standard implementation of these utilities.

This overhaul simplifies our codebase and makes it more reliable, as KLibC has been thoroughly tested and is widely accepted in the developer community. This change also paves the way for implementing more complex functionalities in the future.

Also updated the CMake file to reflect these changes. Now CMake will include the KLibC library when compiling the kernel. A minor adjustment has been made to the kernel.c file to use the new GDT implementation.

In terms of folder structure, moved several source files to a new directory named 'x86_64'. This reflects the specific architecture that these files are intended for, improving code organization.

Updated .idea files to adhere to proposed code standards and cleaned up their unnecessary entries.