Simplify kernel executable source files declaration
This commit streamlines the declaration of source files for kernel executable by removing the explicit listing of 'source/mem/bump_allocator.h' and 'source/mem/bump_allocator.c'. These files are now automatically included in the GLOB_RECURSE command that compiles all .c, .h and .asm files. This results in cleaner, more maintainable code that is easier to manage as the project grows.