"Add Global Descriptor Table (GDT) implementation and its initialization
Updated the kernel initialization sequence to include the initialization of the Global Descriptor Table, a structure important for memory segmentation in the kernel. This includes support for GDT loading and segment refreshing. Updated the CMakeLists.txt for kernel to include correct directory references. Also removed the usage of tabs in the project's idea settings, and corrected a kernel file reference in the build-iso.sh file."