Refactor CMakeLists in RuntimeLib for clarity and efficiency
The CMakeLists.txt file in the RuntimeLib directory was refactored to simplify the code and remove unnecessary flags. The "-fno-lto" and "-fPIE" flags were removed as they are no longer needed. The positioning of include directories has been adjusted for better readability and the compile options for RuntimeLib was changed from PUBLIC to PRIVATE for encapsulation purposes.