Add initial implementation for HostRuntimeLib and RuntimeProtocol
This commit introduces the HostRuntimeLib, which provides a simple implementation of host runtime functions utilizing libc and other OS specific functions, as well as the RuntimeProtocol for data sharing between the host and application.

This is crucial for the functioning of our application loader and necessary for subsequent commands and protocol executions. Additionally, a ProtocolStack functionality was created to aid in the management and access to these protocols, also adjustments made to the CMake lists for the compilation of these new additions.

This foundational work establishes the 'main' function interface and data sharing protocols for host-application interaction, setting stage for more complex functionalities and streamlining future development.