Add key contributing documents and minor setting changes
A new CODE_OF_CONDUCT.md was added to delineate expectations for community behavior and define the enforcement process for any violations. A new CONTRIBUTING.md was created to provide guidelines and expectations for those interested in contributing to this project. This includes sections on how to contribute, report bugs, format commit messages, and pull requests, as well as the project's coding style.

A TIMEOUT setting in the limine.cfg file was changed from 3 to 0.

A line was added in .bazelrc to import user.bazelrc file indicating that user-specific settings can be done in user.bazelrc.

A new README.md was also added to provide an overview of the project, including a simple introduction and instructions on how to build and test the project.

These changes intend to make it clearer for contributors to participate meaningfully in the project, by understanding the community culture and contribution process. The TIMEOUT setting change in limine.cfg is to increase efficiency and the .bazelrc alteration allows for user-specific customization. The README.md serves to give a comprehensive introduction of the project to newcomers.