Update C++ toolchain configuration
This commit updates the configuration for the C++ toolchain. It switches from the older Zig SDK and Hermetic CC Toolchain to the newer LLVM C++ Toolchain for better support and performance. The version has also been upgraded to ensure compatibility with recent updates. Additionally, the commit introduces the use of git_override for linking the repository of the toolchain which enables easier control over the version and source of the toolchain. Some changes were also made to the '.bazelrc' file for improved performance and sandboxing settings.